pyproject.toml
[project]
name = "mcp-gsuite"
version = "0.3.0"
description = "Example MCP server to create a knowledge-base"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
 "beautifulsoup4>=4.12.3",
 "fire>=0.7.0",
 "google-api-python-client>=2.154.0",
 "google-auth>=2.36.0",
 "google-auth-oauthlib>=1.2.2",
 "httplib2>=0.22.0",
 "mcp[cli]>=1.8.1",
 "oauth2client==4.1.3",
 "pydantic-settings>=2.9.1",
 "python-dotenv>=1.0.1",
 "pytz>=2024.2",
 "requests>=2.32.3",
]
[[project.authors]]
name = "Markus Pfundstein"
email = "markus@life-electronic.nl"

[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pyright>=1.1.389",
]

[project.scripts]
mcp-gsuite = "mcp_gsuite:main"
mcp-gsuite-auth = "mcp_gsuite.lib.google_auth_flow:main"
